Lots of small tweaks
