+++ src/modules/launcher/PowerLauncher/Helper/DataWebRequestFactory.cs
@@ -12,23 +12,23 @@ namespace PowerLauncher.Helper
     {
         private class DataWebRequest : WebRequest
         {
-            private readonly Uri m_uri;
+            private readonly Uri _uri;
 
             public DataWebRequest(Uri uri)
             {
-                m_uri = uri;
+                _uri = uri;
             }
 
             public override WebResponse GetResponse()
             {
-                return new DataWebResponse(m_uri);
+                return new DataWebResponse(_uri);
             }
         }
 
         private class DataWebResponse : WebResponse
         {
-            private readonly string m_contentType;
-            private readonly byte[] m_data;
+            private readonly string _contentType;
+            private readonly byte[] _data;
 
             public DataWebResponse(Uri uri)
             {
@@ -36,16 +36,16 @@ namespace PowerLauncher.Helper
 
                 int commaIndex = uriString.IndexOf(',', StringComparison.InvariantCultureIgnoreCase);
                 var headers = uriString.Substring(0, commaIndex).Split(';');
-                m_contentType = headers[0];
+                _contentType = headers[0];
                 string dataString = uriString.Substring(commaIndex + 1);
-                m_data = Convert.FromBase64String(dataString);
+                _data = Convert.FromBase64String(dataString);
             }
 
             public override string ContentType
             {
                 get
                 {
-                    return m_contentType;
+                    return _contentType;
                 }
 
                 set
@@ -58,7 +58,7 @@ namespace PowerLauncher.Helper
             {
                 get
                 {
-                    return m_data.Length;
+                    return _data.Length;
                 }
 
                 set
@@ -69,7 +69,7 @@ namespace PowerLauncher.Helper
 
             public override Stream GetResponseStream()
             {
-                return new MemoryStream(m_data);
+                return new MemoryStream(_data);
             }
         }
