+++ src/modules/launcher/PowerLauncher/Storage/Record.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) Microsoft Corporation
+// The Microsoft Corporation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace PowerLauncher.Storage
+{
+    public class Record
+    {
+        public string Title { get; set; }
+
+        public string SubTitle { get; set; }
+
+        public string PluginID { get; set; }
+    }
+}
